third_party/blink/web_tests/plugins/keyboard-events-expected.txt

Plugin received event: MouseEnter
Plugin received event: MouseMove
Plugin received event: MouseDown
Plugin received event: MouseUp
Plugin received event: RawKeyDown
Plugin received event: Char
Plugin received event: KeyUp
Plugin received event: RawKeyDown
Plugin received event: Char
Plugin received event: KeyUp
Plugin received event: RawKeyDown
Plugin received event: Char
Plugin received event: KeyUp
This test checks if a plugin can receive keyboard events sent from eventSender. This is a test for Bug 34936.
